pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["chg93"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
beautifulsoup4 = "^4.11.1"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
